add fixed block
