+++ rust/src/lib.rs
@@ -1122,33 +1122,6 @@ impl ProtocolVersion {
 }
 
 #[wasm_bindgen]
-#[derive(Clone, Eq, Debug, PartialEq, serde::Serialize, serde::Deserialize, JsonSchema)]
-pub struct TransactionBodies(pub(crate) Vec<TransactionBody>);
-
-impl_to_from!(TransactionBodies);
-
-#[wasm_bindgen]
-impl TransactionBodies {
-    pub fn new() -> Self {
-        Self(Vec::new())
-    }
-
-    pub fn len(&self) -> usize {
-        self.0.len()
-    }
-
-    pub fn get(&self, index: usize) -> TransactionBody {
-        self.0[index].clone()
-    }
-
-    pub fn add(&mut self, elem: &TransactionBody) {
-        self.0.push(elem.clone());
-    }
-}
-
-pub type TransactionIndexes = Vec<TransactionIndex>;
-
-#[wasm_bindgen]
 #[derive(Clone, Debug, Eq, Ord, PartialEq, PartialOrd)]
 pub struct AuxiliaryDataSet(LinkedHashMap<TransactionIndex, AuxiliaryData>);
 
@@ -1213,303 +1186,6 @@ impl JsonSchema for AuxiliaryDataSet {
     }
     fn is_referenceable() -> bool {
         std::collections::BTreeMap::<TransactionIndex, AuxiliaryData>::is_referenceable()
-    }
-}
-
-#[wasm_bindgen]
-#[derive(Clone, Eq, Debug, PartialEq, serde::Serialize, serde::Deserialize, JsonSchema)]
-pub struct Block {
-    header: Header,
-    transaction_bodies: TransactionBodies,
-    transaction_witness_sets: TransactionWitnessSets,
-    auxiliary_data_set: AuxiliaryDataSet,
-    invalid_transactions: TransactionIndexes,
-}
-
-impl_to_from!(Block);
-
-#[wasm_bindgen]
-impl Block {
-    pub fn header(&self) -> Header {
-        self.header.clone()
-    }
-
-    pub fn transaction_bodies(&self) -> TransactionBodies {
-        self.transaction_bodies.clone()
-    }
-
-    pub fn transaction_witness_sets(&self) -> TransactionWitnessSets {
-        self.transaction_witness_sets.clone()
-    }
-
-    pub fn auxiliary_data_set(&self) -> AuxiliaryDataSet {
-        self.auxiliary_data_set.clone()
-    }
-
-    pub fn invalid_transactions(&self) -> TransactionIndexes {
-        self.invalid_transactions.clone()
-    }
-
-    pub fn new(
-        header: &Header,
-        transaction_bodies: &TransactionBodies,
-        transaction_witness_sets: &TransactionWitnessSets,
-        auxiliary_data_set: &AuxiliaryDataSet,
-        invalid_transactions: TransactionIndexes,
-    ) -> Self {
-        Self {
-            header: header.clone(),
-            transaction_bodies: transaction_bodies.clone(),
-            transaction_witness_sets: transaction_witness_sets.clone(),
-            auxiliary_data_set: auxiliary_data_set.clone(),
-            invalid_transactions: invalid_transactions,
-        }
-    }
-}
-
-#[wasm_bindgen]
-#[derive(Clone, Eq, Debug, PartialEq, serde::Serialize, serde::Deserialize, JsonSchema)]
-pub struct Header {
-    header_body: HeaderBody,
-    body_signature: KESSignature,
-}
-
-impl_to_from!(Header);
-
-#[wasm_bindgen]
-impl Header {
-    pub fn header_body(&self) -> HeaderBody {
-        self.header_body.clone()
-    }
-
-    pub fn body_signature(&self) -> KESSignature {
-        self.body_signature.clone()
-    }
-
-    pub fn new(header_body: &HeaderBody, body_signature: &KESSignature) -> Self {
-        Self {
-            header_body: header_body.clone(),
-            body_signature: body_signature.clone(),
-        }
-    }
-}
-
-#[wasm_bindgen]
-#[derive(Clone, Eq, PartialEq, Debug, serde::Serialize, serde::Deserialize, JsonSchema)]
-pub struct OperationalCert {
-    hot_vkey: KESVKey,
-    sequence_number: u32,
-    kes_period: u32,
-    sigma: Ed25519Signature,
-}
-
-impl_to_from!(OperationalCert);
-
-#[wasm_bindgen]
-impl OperationalCert {
-    pub fn hot_vkey(&self) -> KESVKey {
-        self.hot_vkey.clone()
-    }
-
-    pub fn sequence_number(&self) -> u32 {
-        self.sequence_number.clone()
-    }
-
-    pub fn kes_period(&self) -> u32 {
-        self.kes_period.clone()
-    }
-
-    pub fn sigma(&self) -> Ed25519Signature {
-        self.sigma.clone()
-    }
-
-    pub fn new(
-        hot_vkey: &KESVKey,
-        sequence_number: u32,
-        kes_period: u32,
-        sigma: &Ed25519Signature,
-    ) -> Self {
-        Self {
-            hot_vkey: hot_vkey.clone(),
-            sequence_number: sequence_number,
-            kes_period: kes_period,
-            sigma: sigma.clone(),
-        }
-    }
-}
-
-#[derive(Clone, Debug, Eq, PartialEq, serde::Serialize, serde::Deserialize, JsonSchema)]
-pub enum HeaderLeaderCertEnum {
-    NonceAndLeader(VRFCert, VRFCert),
-    VrfResult(VRFCert),
-}
-
-#[wasm_bindgen]
-#[derive(Clone, Eq, PartialEq, Debug, serde::Serialize, serde::Deserialize, JsonSchema)]
-pub struct HeaderBody {
-    block_number: u32,
-    slot: SlotBigNum,
-    prev_hash: Option<BlockHash>,
-    issuer_vkey: Vkey,
-    vrf_vkey: VRFVKey,
-    leader_cert: HeaderLeaderCertEnum,
-    block_body_size: u32,
-    block_body_hash: BlockHash,
-    operational_cert: OperationalCert,
-    protocol_version: ProtocolVersion,
-}
-
-impl_to_from!(HeaderBody);
-
-#[wasm_bindgen]
-impl HeaderBody {
-    pub fn block_number(&self) -> u32 {
-        self.block_number.clone()
-    }
-
-    /// !!! DEPRECATED !!!
-    /// Returns a Slot32 (u32) value in case the underlying original BigNum (u64) value is within the limits.
-    /// Otherwise will just raise an error.
-    #[deprecated(
-        since = "10.1.0",
-        note = "Possible boundary error. Use slot_bignum instead"
-    )]
-    pub fn slot(&self) -> Result<Slot32, JsError> {
-        self.slot.clone().try_into()
-    }
-
-    pub fn slot_bignum(&self) -> SlotBigNum {
-        self.slot.clone()
-    }
-
-    pub fn prev_hash(&self) -> Option<BlockHash> {
-        self.prev_hash.clone()
-    }
-
-    pub fn issuer_vkey(&self) -> Vkey {
-        self.issuer_vkey.clone()
-    }
-
-    pub fn vrf_vkey(&self) -> VRFVKey {
-        self.vrf_vkey.clone()
-    }
-
-    /// If this function returns true, the `.nonce_vrf_or_nothing`
-    /// and the `.leader_vrf_or_nothing` functions will return
-    /// non-empty results
-    pub fn has_nonce_and_leader_vrf(&self) -> bool {
-        match &self.leader_cert {
-            HeaderLeaderCertEnum::NonceAndLeader(_, _) => true,
-            _ => false,
-        }
-    }
-
-    /// Might return nothing in case `.has_nonce_and_leader_vrf` returns false
-    pub fn nonce_vrf_or_nothing(&self) -> Option<VRFCert> {
-        match &self.leader_cert {
-            HeaderLeaderCertEnum::NonceAndLeader(nonce, _) => Some(nonce.clone()),
-            _ => None,
-        }
-    }
-
-    /// Might return nothing in case `.has_nonce_and_leader_vrf` returns false
-    pub fn leader_vrf_or_nothing(&self) -> Option<VRFCert> {
-        match &self.leader_cert {
-            HeaderLeaderCertEnum::NonceAndLeader(_, leader) => Some(leader.clone()),
-            _ => None,
-        }
-    }
-
-    /// If this function returns true, the `.vrf_result_or_nothing`
-    /// function will return a non-empty result
-    pub fn has_vrf_result(&self) -> bool {
-        match &self.leader_cert {
-            HeaderLeaderCertEnum::VrfResult(_) => true,
-            _ => false,
-        }
-    }
-
-    /// Might return nothing in case `.has_vrf_result` returns false
-    pub fn vrf_result_or_nothing(&self) -> Option<VRFCert> {
-        match &self.leader_cert {
-            HeaderLeaderCertEnum::VrfResult(cert) => Some(cert.clone()),
-            _ => None,
-        }
-    }
-
-    pub fn block_body_size(&self) -> u32 {
-        self.block_body_size.clone()
-    }
-
-    pub fn block_body_hash(&self) -> BlockHash {
-        self.block_body_hash.clone()
-    }
-
-    pub fn operational_cert(&self) -> OperationalCert {
-        self.operational_cert.clone()
-    }
-
-    pub fn protocol_version(&self) -> ProtocolVersion {
-        self.protocol_version.clone()
-    }
-
-    /// !!! DEPRECATED !!!
-    /// This constructor uses outdated slot number format.
-    /// Use `.new_headerbody` instead
-    #[deprecated(
-        since = "10.1.0",
-        note = "Underlying value capacity of slot (BigNum u64) bigger then Slot32. Use new_bignum instead."
-    )]
-    pub fn new(
-        block_number: u32,
-        slot: Slot32,
-        prev_hash: Option<BlockHash>,
-        issuer_vkey: &Vkey,
-        vrf_vkey: &VRFVKey,
-        vrf_result: &VRFCert,
-        block_body_size: u32,
-        block_body_hash: &BlockHash,
-        operational_cert: &OperationalCert,
-        protocol_version: &ProtocolVersion,
-    ) -> Self {
-        Self {
-            block_number: block_number,
-            slot: slot.clone().into(),
-            prev_hash: prev_hash.clone(),
-            issuer_vkey: issuer_vkey.clone(),
-            vrf_vkey: vrf_vkey.clone(),
-            leader_cert: HeaderLeaderCertEnum::VrfResult(vrf_result.clone()),
-            block_body_size: block_body_size,
-            block_body_hash: block_body_hash.clone(),
-            operational_cert: operational_cert.clone(),
-            protocol_version: protocol_version.clone(),
-        }
-    }
-
-    pub fn new_headerbody(
-        block_number: u32,
-        slot: &SlotBigNum,
-        prev_hash: Option<BlockHash>,
-        issuer_vkey: &Vkey,
-        vrf_vkey: &VRFVKey,
-        vrf_result: &VRFCert,
-        block_body_size: u32,
-        block_body_hash: &BlockHash,
-        operational_cert: &OperationalCert,
-        protocol_version: &ProtocolVersion,
-    ) -> Self {
-        Self {
-            block_number: block_number,
-            slot: slot.clone(),
-            prev_hash: prev_hash.clone(),
-            issuer_vkey: issuer_vkey.clone(),
-            vrf_vkey: vrf_vkey.clone(),
-            leader_cert: HeaderLeaderCertEnum::VrfResult(vrf_result.clone()),
-            block_body_size: block_body_size,
-            block_body_hash: block_body_hash.clone(),
-            operational_cert: operational_cert.clone(),
-            protocol_version: protocol_version.clone(),
-        }
     }
 }
