+++ rust/src/protocol_types/block/mod.rs
@@ -1,0 +1,29 @@
+mod block;
+pub use block::*;
+
+mod fixed_block;
+pub use fixed_block::*;
+
+mod fixed_tx_body;
+pub use fixed_tx_body::*;
+
+mod header;
+pub use header::*;
+
+mod transaction_bodies;
+pub use transaction_bodies::*;
+
+mod header_body;
+pub use header_body::*;
+
+mod operational_cert;
+pub use operational_cert::*;
+
+mod fixed_versioned_block;
+pub use fixed_versioned_block::*;
+
+mod fixed_transaction_bodies;
+pub use fixed_transaction_bodies::*;
+
+mod versioned_block;
+pub use versioned_block::*;
