+++ rust/src/serialization/block/mod.rs
@@ -1,0 +1,10 @@
+mod block;
+mod header;
+mod operational_cert;
+mod header_body;
+mod transaction_bodies;
+mod fixed_block;
+mod fixed_transaction_body;
+mod fixed_transaction_bodies;
+mod versioned_block;
+mod fixed_versioned_block;
