+++ rust/src/serialization/general.rs
@@ -1300,40 +1300,6 @@ impl DeserializeEmbeddedGroup for ProtocolVersion {
     }
 }
 
-impl cbor_event::se::Serialize for TransactionBodies {
-    fn serialize<'se, W: Write>(
-        &self,
-        serializer: &'se mut Serializer<W>,
-    ) -> cbor_event::Result<&'se mut Serializer<W>> {
-        serializer.write_array(cbor_event::Len::Len(self.0.len() as u64))?;
-        for element in &self.0 {
-            element.serialize(serializer)?;
-        }
-        Ok(serializer)
-    }
-}
-
-impl Deserialize for TransactionBodies {
-    fn deserialize<R: BufRead + Seek>(raw: &mut Deserializer<R>) -> Result<Self, DeserializeError> {
-        let mut arr = Vec::new();
-        (|| -> Result<_, DeserializeError> {
-            let len = raw.array()?;
-            while match len {
-                cbor_event::Len::Len(n) => arr.len() < n as usize,
-                cbor_event::Len::Indefinite => true,
-            } {
-                if is_break_tag(raw, "TransactionBodies")? {
-                    break;
-                }
-                arr.push(TransactionBody::deserialize(raw)?);
-            }
-            Ok(())
-        })()
-        .map_err(|e| e.annotate("TransactionBodies"))?;
-        Ok(Self(arr))
-    }
-}
-
 impl cbor_event::se::Serialize for AuxiliaryDataSet {
     fn serialize<'se, W: Write>(
         &self,
@@ -1373,361 +1339,6 @@ impl Deserialize for AuxiliaryDataSet {
         })()
         .map_err(|e| e.annotate("AuxiliaryDataSet"))?;
         Ok(Self(table))
-    }
-}
-
-impl cbor_event::se::Serialize for Block {
-    fn serialize<'se, W: Write>(
-        &self,
-        serializer: &'se mut Serializer<W>,
-    ) -> cbor_event::Result<&'se mut Serializer<W>> {
-        serializer.write_array(cbor_event::Len::Len(5))?;
-        self.header.serialize(serializer)?;
-        self.transaction_bodies.serialize(serializer)?;
-        self.transaction_witness_sets.serialize(serializer)?;
-        self.auxiliary_data_set.serialize(serializer)?;
-        serializer.write_array(cbor_event::Len::Len(self.invalid_transactions.len() as u64))?;
-        for element in self.invalid_transactions.iter() {
-            element.serialize(serializer)?;
-        }
-        Ok(serializer)
-    }
-}
-
-impl Deserialize for Block {
-    fn deserialize<R: BufRead + Seek>(raw: &mut Deserializer<R>) -> Result<Self, DeserializeError> {
-        (|| -> Result<_, DeserializeError> {
-            let len = raw.array()?;
-            let mut read_len = CBORReadLen::new(len);
-            read_len.read_elems(4)?;
-            let header = (|| -> Result<_, DeserializeError> { Ok(Header::deserialize(raw)?) })()
-                .map_err(|e| e.annotate("header"))?;
-            let transaction_bodies =
-                (|| -> Result<_, DeserializeError> { Ok(TransactionBodies::deserialize(raw)?) })()
-                    .map_err(|e| e.annotate("transaction_bodies"))?;
-            let transaction_witness_sets = (|| -> Result<_, DeserializeError> {
-                Ok(TransactionWitnessSets::deserialize(raw)?)
-            })()
-            .map_err(|e| e.annotate("transaction_witness_sets"))?;
-            let auxiliary_data_set =
-                (|| -> Result<_, DeserializeError> { Ok(AuxiliaryDataSet::deserialize(raw)?) })()
-                    .map_err(|e| e.annotate("auxiliary_data_set"))?;
-            let invalid_present = match len {
-                cbor_event::Len::Indefinite => raw.cbor_type()? == CBORType::Array,
-                cbor_event::Len::Len(4) => false,
-                _ => true,
-            };
-            let invalid_transactions = (|| -> Result<_, DeserializeError> {
-                let mut arr = Vec::new();
-                if invalid_present {
-                    read_len.read_elems(1)?;
-                    let len = raw.array()?;
-                    while match len {
-                        cbor_event::Len::Len(n) => arr.len() < n as usize,
-                        cbor_event::Len::Indefinite => true,
-                    } {
-                        if is_break_tag(raw, "Block.invalid_transactions")? {
-                            break;
-                        }
-                        arr.push(TransactionIndex::deserialize(raw)?);
-                    }
-                }
-                Ok(arr)
-            })()
-            .map_err(|e| e.annotate("invalid_transactions"))?;
-            match len {
-                cbor_event::Len::Len(_) => (),
-                cbor_event::Len::Indefinite => match raw.special()? {
-                    CBORSpecial::Break => (),
-                    _ => return Err(DeserializeFailure::EndingBreakMissing.into()),
-                },
-            }
-            Ok(Block {
-                header,
-                transaction_bodies,
-                transaction_witness_sets,
-                auxiliary_data_set,
-                invalid_transactions,
-            })
-        })()
-        .map_err(|e| e.annotate("Block"))
-    }
-}
-
-impl cbor_event::se::Serialize for Header {
-    fn serialize<'se, W: Write>(
-        &self,
-        serializer: &'se mut Serializer<W>,
-    ) -> cbor_event::Result<&'se mut Serializer<W>> {
-        serializer.write_array(cbor_event::Len::Len(2))?;
-        self.header_body.serialize(serializer)?;
-        self.body_signature.serialize(serializer)?;
-        Ok(serializer)
-    }
-}
-
-impl Deserialize for Header {
-    fn deserialize<R: BufRead + Seek>(raw: &mut Deserializer<R>) -> Result<Self, DeserializeError> {
-        (|| -> Result<_, DeserializeError> {
-            let len = raw.array()?;
-            let ret = Self::deserialize_as_embedded_group(raw, len);
-            match len {
-                cbor_event::Len::Len(_) =>
-                /* TODO: check finite len somewhere */
-                {
-                    ()
-                }
-                cbor_event::Len::Indefinite => match raw.special()? {
-                    CBORSpecial::Break =>
-                    /* it's ok */
-                    {
-                        ()
-                    }
-                    _ => return Err(DeserializeFailure::EndingBreakMissing.into()),
-                },
-            }
-            ret
-        })()
-        .map_err(|e| e.annotate("Header"))
-    }
-}
-
-impl DeserializeEmbeddedGroup for Header {
-    fn deserialize_as_embedded_group<R: BufRead + Seek>(
-        raw: &mut Deserializer<R>,
-        _: cbor_event::Len,
-    ) -> Result<Self, DeserializeError> {
-        let header_body =
-            (|| -> Result<_, DeserializeError> { Ok(HeaderBody::deserialize(raw)?) })()
-                .map_err(|e| e.annotate("header_body"))?;
-        let body_signature =
-            (|| -> Result<_, DeserializeError> { Ok(KESSignature::deserialize(raw)?) })()
-                .map_err(|e| e.annotate("body_signature"))?;
-        Ok(Header {
-            header_body,
-            body_signature,
-        })
-    }
-}
-
-impl cbor_event::se::Serialize for OperationalCert {
-    fn serialize<'se, W: Write>(
-        &self,
-        serializer: &'se mut Serializer<W>,
-    ) -> cbor_event::Result<&'se mut Serializer<W>> {
-        serializer.write_array(cbor_event::Len::Len(4))?;
-        self.serialize_as_embedded_group(serializer)
-    }
-}
-
-impl SerializeEmbeddedGroup for OperationalCert {
-    fn serialize_as_embedded_group<'se, W: Write>(
-        &self,
-        serializer: &'se mut Serializer<W>,
-    ) -> cbor_event::Result<&'se mut Serializer<W>> {
-        self.hot_vkey.serialize(serializer)?;
-        self.sequence_number.serialize(serializer)?;
-        self.kes_period.serialize(serializer)?;
-        self.sigma.serialize(serializer)?;
-        Ok(serializer)
-    }
-}
-
-impl Deserialize for OperationalCert {
-    fn deserialize<R: BufRead + Seek>(raw: &mut Deserializer<R>) -> Result<Self, DeserializeError> {
-        (|| -> Result<_, DeserializeError> {
-            let len = raw.array()?;
-            let ret = Self::deserialize_as_embedded_group(raw, len);
-            match len {
-                cbor_event::Len::Len(_) =>
-                /* TODO: check finite len somewhere */
-                {
-                    ()
-                }
-                cbor_event::Len::Indefinite => match raw.special()? {
-                    CBORSpecial::Break =>
-                    /* it's ok */
-                    {
-                        ()
-                    }
-                    _ => return Err(DeserializeFailure::EndingBreakMissing.into()),
-                },
-            }
-            ret
-        })()
-        .map_err(|e| e.annotate("OperationalCert"))
-    }
-}
-
-impl DeserializeEmbeddedGroup for OperationalCert {
-    fn deserialize_as_embedded_group<R: BufRead + Seek>(
-        raw: &mut Deserializer<R>,
-        _: cbor_event::Len,
-    ) -> Result<Self, DeserializeError> {
-        let hot_vkey = (|| -> Result<_, DeserializeError> { Ok(KESVKey::deserialize(raw)?) })()
-            .map_err(|e| e.annotate("hot_vkey"))?;
-        let sequence_number = (|| -> Result<_, DeserializeError> { Ok(u32::deserialize(raw)?) })()
-            .map_err(|e| e.annotate("sequence_number"))?;
-        let kes_period = (|| -> Result<_, DeserializeError> { Ok(u32::deserialize(raw)?) })()
-            .map_err(|e| e.annotate("kes_period"))?;
-        let sigma =
-            (|| -> Result<_, DeserializeError> { Ok(Ed25519Signature::deserialize(raw)?) })()
-                .map_err(|e| e.annotate("sigma"))?;
-        Ok(OperationalCert {
-            hot_vkey,
-            sequence_number,
-            kes_period,
-            sigma,
-        })
-    }
-}
-
-impl cbor_event::se::Serialize for HeaderBody {
-    fn serialize<'se, W: Write>(
-        &self,
-        serializer: &'se mut Serializer<W>,
-    ) -> cbor_event::Result<&'se mut Serializer<W>> {
-        serializer.write_array(cbor_event::Len::Len(15))?;
-        self.block_number.serialize(serializer)?;
-        self.slot.serialize(serializer)?;
-        match &self.prev_hash {
-            Some(x) => x.serialize(serializer),
-            None => serializer.write_special(CBORSpecial::Null),
-        }?;
-        self.issuer_vkey.serialize(serializer)?;
-        self.vrf_vkey.serialize(serializer)?;
-        match &self.leader_cert {
-            HeaderLeaderCertEnum::NonceAndLeader(nonce_vrf, leader_vrf) => {
-                nonce_vrf.serialize(serializer)?;
-                leader_vrf.serialize(serializer)?;
-            }
-            HeaderLeaderCertEnum::VrfResult(vrf_cert) => {
-                vrf_cert.serialize(serializer)?;
-            }
-        }
-        self.block_body_size.serialize(serializer)?;
-        self.block_body_hash.serialize(serializer)?;
-        self.operational_cert
-            .serialize_as_embedded_group(serializer)?;
-        self.protocol_version
-            .serialize_as_embedded_group(serializer)?;
-        Ok(serializer)
-    }
-}
-
-impl Deserialize for HeaderBody {
-    fn deserialize<R: BufRead + Seek>(raw: &mut Deserializer<R>) -> Result<Self, DeserializeError> {
-        (|| -> Result<_, DeserializeError> {
-            let len = raw.array()?;
-            let ret = Self::deserialize_as_embedded_group(raw, len);
-            match len {
-                cbor_event::Len::Len(_) =>
-                /* TODO: check finite len somewhere */
-                {
-                    ()
-                }
-                cbor_event::Len::Indefinite => match raw.special()? {
-                    CBORSpecial::Break =>
-                    /* it's ok */
-                    {
-                        ()
-                    }
-                    _ => return Err(DeserializeFailure::EndingBreakMissing.into()),
-                },
-            }
-            ret
-        })()
-        .map_err(|e| e.annotate("HeaderBody"))
-    }
-}
-
-impl DeserializeEmbeddedGroup for HeaderBody {
-    fn deserialize_as_embedded_group<R: BufRead + Seek>(
-        raw: &mut Deserializer<R>,
-        len: cbor_event::Len,
-    ) -> Result<Self, DeserializeError> {
-        let block_number = (|| -> Result<_, DeserializeError> { Ok(u32::deserialize(raw)?) })()
-            .map_err(|e| e.annotate("block_number"))?;
-        let slot = (|| -> Result<_, DeserializeError> { Ok(SlotBigNum::deserialize(raw)?) })()
-            .map_err(|e| e.annotate("slot"))?;
-        let prev_hash = (|| -> Result<_, DeserializeError> {
-            Ok(match raw.cbor_type()? != CBORType::Special {
-                true => Some(BlockHash::deserialize(raw)?),
-                false => {
-                    if raw.special()? != CBORSpecial::Null {
-                        return Err(DeserializeFailure::ExpectedNull.into());
-                    }
-                    None
-                }
-            })
-        })()
-        .map_err(|e| e.annotate("prev_hash"))?;
-        let issuer_vkey = (|| -> Result<_, DeserializeError> { Ok(Vkey::deserialize(raw)?) })()
-            .map_err(|e| e.annotate("issuer_vkey"))?;
-        let vrf_vkey = (|| -> Result<_, DeserializeError> { Ok(VRFVKey::deserialize(raw)?) })()
-            .map_err(|e| e.annotate("vrf_vkey"))?;
-        let leader_cert = {
-            // NONCE VFR CERT, first of two certs
-            // or a single VRF RESULT CERT
-            // depending on the protocol version
-            let first_vrf_cert =
-                (|| -> Result<_, DeserializeError> { Ok(VRFCert::deserialize(raw)?) })()
-                    .map_err(|e| e.annotate("nonce_vrf"))?;
-            let cbor_type: cbor_event::Type = raw.cbor_type()?;
-            match cbor_type {
-                cbor_event::Type::Array => {
-                    // Legacy format, reading the second VRF cert
-                    let leader_vrf =
-                        (|| -> Result<_, DeserializeError> { Ok(VRFCert::deserialize(raw)?) })()
-                            .map_err(|e| e.annotate("leader_vrf"))?;
-                    HeaderLeaderCertEnum::NonceAndLeader(first_vrf_cert, leader_vrf)
-                }
-                cbor_event::Type::UnsignedInteger => {
-                    // New format, no second VRF cert is present
-                    HeaderLeaderCertEnum::VrfResult(first_vrf_cert)
-                }
-                t => {
-                    return Err(DeserializeError::new(
-                        "HeaderBody.leader_cert",
-                        DeserializeFailure::UnexpectedKeyType(t),
-                    ))
-                }
-            }
-        };
-        let block_body_size = (|| -> Result<_, DeserializeError> { Ok(u32::deserialize(raw)?) })()
-            .map_err(|e| e.annotate("block_body_size"))?;
-        let block_body_hash =
-            (|| -> Result<_, DeserializeError> { Ok(BlockHash::deserialize(raw)?) })()
-                .map_err(|e| e.annotate("block_body_hash"))?;
-
-        let operational_cert = (|| -> Result<_, DeserializeError> {
-            if raw.cbor_type()? == CBORType::Array {
-                Ok(OperationalCert::deserialize(raw)?)
-            } else {
-                Ok(OperationalCert::deserialize_as_embedded_group(raw, len)?)
-            }
-        })()
-        .map_err(|e| e.annotate("operational_cert"))?;
-        let protocol_version = (|| -> Result<_, DeserializeError> {
-            if raw.cbor_type()? == CBORType::Array {
-                Ok(ProtocolVersion::deserialize(raw)?)
-            } else {
-                Ok(ProtocolVersion::deserialize_as_embedded_group(raw, len)?)
-            }
-        })()
-        .map_err(|e| e.annotate("protocol_version"))?;
-        Ok(HeaderBody {
-            block_number,
-            slot,
-            prev_hash,
-            issuer_vkey,
-            vrf_vkey,
-            leader_cert,
-            block_body_size,
-            block_body_hash,
-            operational_cert,
-            protocol_version,
-        })
     }
 }
