+++ rust/src/serialization/mod.rs
@@ -25,6 +25,6 @@ mod plutus;
 mod native_script;
 mod native_scripts;
 mod numeric;
-mod versioned_block;
 mod script_ref;
 mod tx_input;
+mod block;
